Removing board.breakouts and resolving IMU.Drivers at runtime
