+++ lib/accelerometer.js
@@ -43,7 +43,8 @@ var Controllers = {
   MPU6050: {
     initialize: {
       value: function(opts, dataHandler) {
-        var driver = Board.Breakouts.get(this.board, "MPU-6050", opts),
+        var IMU = require("../lib/imu");
+        var driver = IMU.Drivers.get(this.board, "MPU6050", opts),
           state = priv.get(this);
 
         state.maxValue = opts.maxValue || 17300;
