+++ lib/board.js
@@ -8,7 +8,6 @@ var __ = require("../lib/fn.js");
 var Repl = require("../lib/repl.js");
 var Options = require("../lib/board.options.js");
 var Pins = require("../lib/board.pins.js");
-var Breakouts = require("../lib/board.breakouts.js");
 // var temporal = require("temporal"),
 var IO;
 
@@ -830,8 +829,6 @@ if (IS_TEST_MODE) {
     Serial: Serial
   };
 }
-
-Board.Breakouts = Breakouts;
 
 module.exports = Board;
