+++ lib/gyro.js
@@ -61,8 +61,9 @@ var Controllers = {
   MPU6050: {
     initialize: {
       value: function(opts, dataHandler) {
+        var IMU = require("../lib/imu");
         var state = priv.get(this),
-          driver = Board.Breakouts.get(this.board, "MPU-6050", opts);
+          driver = IMU.Drivers.get(this.board, "MPU-6050", opts);
 
         state.sensitivity = opts.sensitivity || 131;
