+++ lib/temperature.js
@@ -140,7 +140,8 @@ var Controllers = {
   MPU6050: {
     initialize: {
       value: function(opts, dataHandler) {
-        var driver = Board.Breakouts.get(this.board, "MPU6050", opts);
+        var IMU = require("../lib/imu");
+        var driver = IMU.Drivers.get(this.board, "MPU6050", opts);
         driver.on("data", function(err, data) {
           dataHandler.call(this, data.temperature);
         }.bind(this));
